backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
python-dotenv==0.19.0
openai==1.3.0
elevenlabs==0.2.24
python-magic==0.4.27
